pyproject.toml
[project]
name = "summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.0",
    "discord-py>=2.3.2",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.3",
    "python-dotenv>=1.0.1",
    "ruff>=0.11.4",
    "requests>=2.32.3",
    "langchain-community>=0.3.21",
    "google-cloud-aiplatform>=1.88.0",
    "feedparser>=6.0.11",
    "google-genai>=1.31.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "ruff>=0.11.4",
]

[tool.ruff.lint]
select = ["RUF", "I", "UP", "SIM", "FURB", "PIE", "PERF", "C4", "B", "PL", "S"]
ignore = ["RUF001", "RUF002", "B008", "S101"]


